FPU・内部発振を有効化

+++ firmware/libs/CMSIS/include/stm32f4xx.h
@@ -119,20 +119,20 @@
   */           
 
 #if !defined  (HSE_VALUE) 
-  #define HSE_VALUE    ((uint32_t)25000000) /*!< Value of the External oscillator in Hz */
+  #define HSE_VALUE    (25000000) /*!< Value of the External oscillator in Hz */
   
 #endif /* HSE_VALUE */
 
-/**
+ /**
  * @brief In the following line adjust the External High Speed oscillator (HSE) Startup 
    Timeout value 
    */
 #if !defined  (HSE_STARTUP_TIMEOUT) 
-  #define HSE_STARTUP_TIMEOUT    ((uint16_t)0x05000)   /*!< Time out for HSE start up */
+  #define HSE_STARTUP_TIMEOUT    (0x05000)   /*!< Time out for HSE start up */
 #endif /* HSE_STARTUP_TIMEOUT */   
 
 #if !defined  (HSI_VALUE)   
-  #define HSI_VALUE    ((uint32_t)16000000) /*!< Value of the Internal oscillator in Hz*/
+  #define HSI_VALUE    (16000000) /*!< Value of the Internal oscillator in Hz*/
 #endif /* HSI_VALUE */   
 
 /**
